+++ test/install_generator_test.rb
@@ -23,6 +23,7 @@ class InstallGeneratorTest < Minitest::Test
     assert_includes initializer, "additional_profile_attributes"
     assert_includes initializer, "require_password_confirmation"
     assert_includes initializer, "config.login_title"
+    assert_includes initializer, "primary_login_type"
     assert_includes initializer, "omniauth_providers"
     assert_includes initializer, "omniauth_create_account"
     assert_includes initializer, "recording_studio_user/omniauth_callbacks"